requirements.txt
dnslib==0.9.23
gevent==22.10.2
docker-py==1.10.6
anyconfig==0.13.0
